Refactor: 오늘의 피드 노출 기준 재설정
