+++ STORIX-Batch/src/main/java/com/storix/StorixBatchApplication.java
@@ -1,15 +1,22 @@
-package com.storix.batch;
+package com.storix;
 
 import jakarta.annotation.PostConstruct;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.FilterType;
 import org.springframework.scheduling.annotation.EnableScheduling;
+import com.storix.infrastructure.config.SecurityConfig;
 
 import java.util.TimeZone;
 
 @EnableScheduling
-@SpringBootApplication(scanBasePackages = "com.storix")
+@SpringBootApplication
+@ComponentScan(
+        basePackages = "com.storix",
+        excludeFilters = @ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE, classes = SecurityConfig.class)
+)
 public class StorixBatchApplication {
 
     @PostConstruct
